Create a rake task to pre-create our queues
Google Pub/Sub requires all queues to be pre-created before we can publish messages to them.

Furthermore, if you publish to a topic without a subscription then the message is silently dropped. Because of this we also pre-create a subscription and remove the default death-timer of 31 days — this way we'll always have a subscription for messages to go to!