Make failure more clear when emulator isnt there
Google Pub/Sub has no inherent timeout, but will hang hard (forever) if emulator isn't running when we try to connect to it (e.g. run the tests). This is not very helpful, so we attempt to connect by ourselves first,
and then bail hard if the emulator isn't available.